requirements.txt
Flask == 1.1.2
httpie == 2.3.0
flask_sqlalchemy == 2.4.4
psycopg2 == 2.8.6
python-dateutil ==2.8.1